requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
python-multipart
psycopg2-binary
